requirements.txt
ipykernel
yfinance
pandas
numpy
streamlit
plotly.express